fix: moved all logic to hook
